Move CodecMode from amqp-message to amqp-publisher
CodecMode controls publish-side type narrowing, not message structure.
Colocate it with the related Body/PlainBody types.
