+++ src/amqp-publisher.ts
@@ -1,6 +1,8 @@
 import type { AMQPProperties } from "./amqp-properties.js"
 import type { AMQPSession } from "./amqp-session.js"
-import type { CodecMode } from "./amqp-message.js"
+
+/** Controls whether session methods accept rich types or only raw bytes. */
+export type CodecMode = "plain" | "codec"
 
 export type PlainBody = string | Uint8Array | ArrayBuffer | Buffer | null
 type Serializable = PlainBody | number | boolean | Record<string, unknown> | unknown[]
