+++ src/amqp-queue.ts
@@ -6,8 +6,7 @@ import { AMQPSubscription, AMQPGeneratorSubscription } from "./amqp-subscription
 import type { ConsumerDefinition } from "./amqp-subscription.js"
 import type { AMQPSession } from "./amqp-session.js"
 import { publishConfirmed, publishNoConfirm } from "./amqp-publisher.js"
-import type { Body } from "./amqp-publisher.js"
-import type { CodecMode } from "./amqp-message.js"
+import type { Body, CodecMode } from "./amqp-publisher.js"
 import type { AMQPCodecRegistry } from "./amqp-codec-registry.js"
 
 /**
