+++ src/index.ts
@@ -22,4 +22,4 @@ export type { Logger } from "./types.js"
 export type { AMQPWebSocketInit } from "./amqp-websocket-client.js"
 export { AMQPCodecRegistry } from "./amqp-codec-registry.js"
 export type { AMQPParser, AMQPCoder } from "./amqp-codec-registry.js"
-export type { Body } from "./amqp-publisher.js"
+export type { Body, CodecMode } from "./amqp-publisher.js"
